requirements.txt
pytest==8.3.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.8.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
cryptography==3.4.8
exceptiongroup==1.2.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
pytz==2022.1
Werkzeug==3.0.4














